# alignment_openapi.yaml
openapi: 3.0.3
info:
  title: ATOM AI Backend API
  version: 1.0.0
  description: API contract for ATOM AI Workspace
paths:
...
  /api/auth/invite/activate:
    post:
      operationId: authInviteActivate
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteActivate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InviteActivate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InviteActivate'
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteActivateResponse'
          description: ''
  /api/auth/login:
    post:
      operationId: authLogin
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLogin'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AuthLogin'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AuthLogin'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: ''
  /api/auth/logout:
    post:
      operationId: authLogout
      tags:
      - auth
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '204':
          description: Logged out
  /api/auth/session:
    get:
      operationId: authSession
      tags:
      - auth
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: ''
...
  /api/projects/{id}/archive:
    post:
      operationId: projectArchive
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
...
  /api/projects/{id}/members/{member_id}:
    patch:
      operationId: projectMemberUpdate
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: member_id
        schema:
          type: integer
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProjectMemberUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProjectMemberUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProjectMemberUpdate'
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
          description: ''
    delete:
      operationId: projectMemberDelete
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: path
        name: member_id
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '204':
          description: Project member deleted
  /api/projects/{id}/restore:
    post:
      operationId: projectRestore
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - projects
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
...
components:
  schemas:
...
    AuthLogin:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required:
      - password
...
    InviteActivate:
      type: object
      properties:
        invite_token:
          type: string
        token:
          type: string
        password:
          type: string
    InviteActivateResponse:
      type: object
      properties:
        detail:
          type: string
        invite_token:
          type: string
      required:
      - detail
      - invite_token
...
    PatchedProjectMemberUpdate:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/ProjectMemberRoleEnum'
        is_active:
          type: boolean
...
    Project:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        organization_id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatusEnum'
        created_by_id:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - created_by_id
      - id
      - name
      - organization_id
      - updated_at
...
    ProjectMember:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        project:
          type: integer
        user_id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
        email:
          type: string
          readOnly: true
        first_name:
          type: string
          readOnly: true
        last_name:
          type: string
          readOnly: true
        role:
          $ref: '#/components/schemas/ProjectMemberRoleEnum'
        is_active:
          type: boolean
        joined_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - email
      - first_name
      - id
      - joined_at
      - last_name
      - project
      - user_id
      - username
...
    ProjectMemberRoleEnum:
      enum:
      - owner
      - editor
      - viewer
      type: string
      description: |-
        * `owner` - Owner
        * `editor` - Editor
        * `viewer` - Viewer
    ProjectStatusEnum:
      enum:
      - active
      - on_hold
      - completed
      - archived
      type: string
      description: |-
        * `active` - Active
        * `on_hold` - On Hold
        * `completed` - Completed
        * `archived` - Archived
...
    SessionPayload:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/SessionUser'
      required:
      - token
      - user
    SessionResponse:
      type: object
      properties:
        session:
          $ref: '#/components/schemas/SessionPayload'
      required:
      - session
    SessionUser:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        email:
          type: string
          format: email
          title: Адрес электронной почты
          maxLength: 254
        role:
          type: string
          readOnly: true
        department:
          type: string
          readOnly: true
      required:
      - department
      - id
      - name
      - role
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid